# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Digital Banking Simulator API
  version: 2.0.0
  description: Educational simulator only. No real financial transactions.
servers:
  - url: http://localhost:8080/api/v1
    description: Local backend via proxy
  - url: /api/v1
    description: Same-origin deployment
security:
  - bearerAuth: []
paths:
  /auth/register/send-otp:
    post:
      operationId: sendRegistrationOtp
      tags: [Auth]
      security: []
      description: Send 6-digit registration OTP to phone number before submitting registration.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SendOtpRequest' }
      responses:
        '200':
          description: OTP dispatched successfully
          content: { application/json: { schema: { $ref: '#/components/schemas/SendOtpResponse' } } }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '409': { $ref: '#/components/responses/ProblemResponse' }
        '429': { $ref: '#/components/responses/ProblemResponse' }
  /auth/register:
    post:
      operationId: registerCustomer
      tags: [Auth]
      security: []
      description: Self-service registration with verified phone OTP. Creates Customer and active default checking account immediately.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Customer and account created successfully
          content: { application/json: { schema: { $ref: '#/components/schemas/RegistrationResponse' } } }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '409': { $ref: '#/components/responses/ProblemResponse' }
        '429': { $ref: '#/components/responses/ProblemResponse' }
  /auth/login:
    post:
      operationId: login
      tags: [Auth]
      security: []
      description: Authenticate with registered phone number and password. Returns access token, HttpOnly refresh cookie, and isPinSet flag.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated; refresh token is HttpOnly cookie
          headers:
            Set-Cookie: { schema: { type: string }, description: HttpOnly refresh_token cookie }
          content: { application/json: { schema: { $ref: '#/components/schemas/LoginResponse' } } }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '429': { $ref: '#/components/responses/ProblemResponse' }
...
  /auth/recover/initiate:
    post:
      operationId: initiateAccountRecovery
      tags: [Auth]
      security: []
      description: >-
        Request recovery OTP via chosen channel (SMS to registered phone or EMAIL to registered email).
        Anti-enumeration: always returns the same 200 body whether or not the identifier is registered;
        OTP is dispatched only when a matching account exists. A new request invalidates the previous recovery OTP and any active reset token for that account.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecoverInitiateRequest' }
      responses:
        '200':
          description: Generic response, identical for registered and unregistered identifiers
          content: { application/json: { schema: { $ref: '#/components/schemas/RecoverInitiateResponse' } } }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '429': { $ref: '#/components/responses/ProblemResponse' }
  /auth/recover/verify:
    post:
      operationId: verifyAccountRecoveryOtp
      tags: [Auth]
      security: []
      description: >-
        Verify the single-use 6-digit OTP (TTL 120s) for the chosen SMS or EMAIL channel.
        On success, consume the OTP and return a one-time reset token valid for 300 seconds.
        Wrong, expired, used OTP and unknown identifier return the same 400 OTP_INVALID.
        Each independent IP and identifier rate-limit bucket allows 5 requests per 300 seconds.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecoverVerifyRequest' }
      responses:
        '200':
          description: OTP verified; reset token issued
          content: { application/json: { schema: { $ref: '#/components/schemas/RecoverVerifyResponse' } } }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '429': { $ref: '#/components/responses/ProblemResponse' }
  /auth/recover/confirm:
    post:
      operationId: confirmAccountRecovery
      tags: [Auth]
      security: []
      description: >-
        Accept only the one-time reset token obtained from /auth/recover/verify and a new password.
        Atomically consume the token, save the new password and revoke all refresh tokens of the user.
        Missing, invalid, expired or used token returns 400 RECOVERY_TOKEN_INVALID.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecoverConfirmRequest' }
      responses:
        '200':
          description: Password reset successfully and active sessions revoked
          content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '429': { $ref: '#/components/responses/ProblemResponse' }
  /customers/me:
    get:
      operationId: getCurrentCustomer
      tags: [Customer]
      description: Get caller profile including phone, email, and isPinSet flag.
      responses:
        '200': { description: Own profile, content: { application/json: { schema: { $ref: '#/components/schemas/CustomerProfile' } } } }
        '401': { $ref: '#/components/responses/ProblemResponse' }
  /customers/me/pin/setup:
    post:
      operationId: setupPin
      tags: [Customer]
      description: Mandatory first-login transaction PIN configuration (6 digits). Required before any transfer can be initiated.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetupPinRequest' }
      responses:
        '200': { description: Transaction PIN configured successfully }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '409': { $ref: '#/components/responses/ProblemResponse' }
  /customers/me/pin/change:
    post:
      operationId: changePin
      tags: [Customer]
      description: Change transaction PIN by providing valid current PIN.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePinRequest' }
      responses:
        '200': { description: Transaction PIN changed successfully }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '403': { $ref: '#/components/responses/ProblemResponse' }
  /customers/me/pin/forgot/initiate:
    post:
      operationId: initiateForgotPin
      tags: [Customer]
      description: Send reset OTP to customer registered phone number to reset forgotten PIN.
      responses:
        '200':
          description: Reset OTP sent to registered phone number
          content: { application/json: { schema: { $ref: '#/components/schemas/SendOtpResponse' } } }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '429': { $ref: '#/components/responses/ProblemResponse' }
  /customers/me/pin/forgot/confirm:
    post:
      operationId: confirmForgotPin
      tags: [Customer]
      description: Validate phone OTP and set new transaction PIN.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgotPinConfirmRequest' }
      responses:
        '200': { description: Transaction PIN reset successfully }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
  /operator/customers/send-otp:
    post:
      operationId: operatorSendCustomerOtp
      tags: [Operator]
      x-required-roles: [OPERATOR, ADMIN]
      description: Operator initiates customer creation at counter by triggering OTP to customer present phone number.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperatorSendOtpRequest' }
      responses:
        '200':
          description: OTP sent to customer phone at counter
          content: { application/json: { schema: { $ref: '#/components/schemas/SendOtpResponse' } } }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '403': { $ref: '#/components/responses/ProblemResponse' }
        '409': { $ref: '#/components/responses/ProblemResponse' }
        '429': { $ref: '#/components/responses/ProblemResponse' }
  /operator/customers:
    post:
      operationId: operatorCreateCustomer
      tags: [Operator]
      x-required-roles: [OPERATOR, ADMIN]
      description: Operator provisions customer and active account at counter with verified OTP and initial password. Customer must configure PIN on first login.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperatorCreateCustomerRequest' }
      responses:
        '201':
          description: Customer and account created at counter
          content: { application/json: { schema: { $ref: '#/components/schemas/RegistrationResponse' } } }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '403': { $ref: '#/components/responses/ProblemResponse' }
        '409': { $ref: '#/components/responses/ProblemResponse' }
    get:
      operationId: operatorLookupCustomer
      tags: [Operator]
      x-required-roles: [OPERATOR, ADMIN]
      description: >-
        Exact-match customer lookup at counter. Exactly one of phone or email is required; unfiltered listing,
        partial match and wildcard are rejected with 400 VALIDATION_ERROR. Every lookup is audited and rate limited.
      parameters:
        - { name: phone, in: query, schema: { $ref: '#/components/schemas/PhoneNumber' } }
        - { name: email, in: query, schema: { $ref: '#/components/schemas/EmailAddress' } }
      responses:
        '200': { description: Matching customer with accounts, content: { application/json: { schema: { $ref: '#/components/schemas/OperatorCustomerView' } } } }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '403': { $ref: '#/components/responses/ProblemResponse' }
        '404': { $ref: '#/components/responses/ProblemResponse' }
        '429': { $ref: '#/components/responses/ProblemResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ProblemResponse:
      description: RFC 9457-compatible error
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  schemas:
    Uuid: { type: string, format: uuid }
    Instant: { type: string, format: date-time, description: RFC 3339 UTC timestamp emitted with Z suffix }
    PhoneNumber: { type: string, pattern: '^0[3-9][0-9]{8}$', examples: ['0912345678'], description: Unique 10-digit Vietnam phone number }
    EmailAddress: { type: string, format: email, maxLength: 254, description: Unique email address }
    MoneyAmount: { type: string, pattern: '^(0|[1-9][0-9]{0,14})$', examples: ['50000'], description: Simulated VND, integer only (scale 0) }
    Currency: { type: string, enum: [VND] }
    AccountStatus: { type: string, enum: [ACTIVE, BLOCKED, CLOSED] }
    Role: { type: string, enum: [CUSTOMER, OPERATOR, AUDITOR, ADMIN] }
    RecoveryChannel: { type: string, enum: [SMS, EMAIL] }
    SendOtpRequest:
      type: object
      additionalProperties: false
      required: [phone, purpose]
      properties:
        phone: { $ref: '#/components/schemas/PhoneNumber' }
        purpose: { type: string, enum: [REGISTRATION, OPERATOR_CREATE_CUSTOMER] }
    SendOtpResponse:
      type: object
      required: [phone, expiresInSeconds, message]
      properties:
        phone: { type: string }
        expiresInSeconds: { type: integer, example: 120 }
        message: { type: string, example: "OTP sent successfully" }
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [phone, email, password, fullName, otp]
      properties:
        phone: { $ref: '#/components/schemas/PhoneNumber' }
        email: { $ref: '#/components/schemas/EmailAddress' }
        password: { type: string, minLength: 12, maxLength: 128, writeOnly: true }
        fullName: { type: string, minLength: 1, maxLength: 120 }
        otp: { type: string, pattern: '^[0-9]{6}$', description: 6-digit OTP verified via SMS }
        address: { type: string, maxLength: 300 }
    RegistrationResponse:
      type: object
      required: [customerId, phone, email, fullName, account, createdAt]
      properties:
        customerId: { $ref: '#/components/schemas/Uuid' }
        phone: { $ref: '#/components/schemas/PhoneNumber' }
        email: { $ref: '#/components/schemas/EmailAddress' }
        fullName: { type: string }
        account: { $ref: '#/components/schemas/Account' }
        createdAt: { $ref: '#/components/schemas/Instant' }
    OperatorSendOtpRequest:
      type: object
      additionalProperties: false
      required: [phone]
      properties:
        phone: { $ref: '#/components/schemas/PhoneNumber' }
    OperatorCreateCustomerRequest:
      type: object
      additionalProperties: false
      required: [phone, email, fullName, initialPassword, otp]
      properties:
        phone: { $ref: '#/components/schemas/PhoneNumber' }
        email: { $ref: '#/components/schemas/EmailAddress' }
        fullName: { type: string, minLength: 1, maxLength: 120 }
        initialPassword: { type: string, minLength: 12, maxLength: 128, writeOnly: true, description: Initial password assigned by operator }
        otp: { type: string, pattern: '^[0-9]{6}$', description: 6-digit OTP confirmed by customer at counter }
        address: { type: string, maxLength: 300 }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [phone, password]
      properties:
        phone: { $ref: '#/components/schemas/PhoneNumber' }
        password: { type: string, minLength: 1, maxLength: 128, writeOnly: true }
    UserSummary:
      type: object
      required: [userId, customerId, displayName, phone, email, roles, isPinSet]
      properties:
        userId: { $ref: '#/components/schemas/Uuid' }
        customerId: { $ref: '#/components/schemas/Uuid' }
        displayName: { type: string }
        phone: { $ref: '#/components/schemas/PhoneNumber' }
        email: { $ref: '#/components/schemas/EmailAddress' }
        roles: { type: array, items: { $ref: '#/components/schemas/Role' } }
        isPinSet: { type: boolean, description: Indicates if transaction PIN is already configured }
    AccessTokenResponse:
      type: object
      required: [accessToken, tokenType, expiresIn]
      properties:
        accessToken: { type: string }
        tokenType: { type: string, enum: [Bearer] }
        expiresIn: { type: integer, minimum: 1 }
    LoginResponse:
      allOf:
        - { $ref: '#/components/schemas/AccessTokenResponse' }
        - type: object
          required: [user]
          properties: { user: { $ref: '#/components/schemas/UserSummary' } }
...
    CustomerProfile:
      type: object
      required: [customerId, fullName, phone, email, isPinSet, createdAt]
      properties:
        customerId: { $ref: '#/components/schemas/Uuid' }
        fullName: { type: string }
        phone: { $ref: '#/components/schemas/PhoneNumber' }
        email: { $ref: '#/components/schemas/EmailAddress' }
        isPinSet: { type: boolean }
        createdAt: { $ref: '#/components/schemas/Instant' }
    RecoverInitiateRequest:
      type: object
      additionalProperties: false
      required: [identifier, channel]
      properties:
        identifier: { type: string, description: Registered phone number or email address }
        channel: { $ref: '#/components/schemas/RecoveryChannel' }
    RecoverInitiateResponse:
      type: object
      required: [identifier, channel, expiresInSeconds, message]
      properties:
        identifier: { type: string }
        channel: { $ref: '#/components/schemas/RecoveryChannel' }
        expiresInSeconds: { type: integer, example: 120 }
        message: { type: string, example: "If the information is registered, an OTP has been sent to the selected channel." }
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string, example: "Password reset successfully and active sessions revoked" }
    RecoverVerifyRequest:
      type: object
      additionalProperties: false
      required: [identifier, channel, otp]
      properties:
        identifier: { type: string }
        channel: { $ref: '#/components/schemas/RecoveryChannel' }
        otp: { type: string, pattern: '^[0-9]{6}$' }
    RecoverVerifyResponse:
      type: object
      required: [resetToken, expiresInSeconds]
      properties:
        resetToken: { type: string, description: 'Opaque one-time recovery proof; keep in memory only and never put in a URL or log' }
        expiresInSeconds: { type: integer, example: 300 }
    RecoverConfirmRequest:
      type: object
      additionalProperties: false
      required: [resetToken, newPassword]
      properties:
        resetToken: { type: string, description: 'One-time token returned by recovery verify' }
        newPassword: { type: string, minLength: 12, maxLength: 128, writeOnly: true }
    SetupPinRequest:
      type: object
      additionalProperties: false
      required: [pin, confirmPin]
      properties:
        pin: { type: string, pattern: '^[0-9]{6}$', writeOnly: true, description: 6-digit transaction PIN }
        confirmPin: { type: string, pattern: '^[0-9]{6}$', writeOnly: true }
    ChangePinRequest:
      type: object
      additionalProperties: false
      required: [currentPin, newPin, confirmNewPin]
      properties:
        currentPin: { type: string, pattern: '^[0-9]{6}$', writeOnly: true }
        newPin: { type: string, pattern: '^[0-9]{6}$', writeOnly: true }
        confirmNewPin: { type: string, pattern: '^[0-9]{6}$', writeOnly: true }
    ForgotPinConfirmRequest:
      type: object
      additionalProperties: false
      required: [otp, newPin, confirmNewPin]
      properties:
        otp: { type: string, pattern: '^[0-9]{6}$', description: 6-digit OTP sent to registered phone }
        newPin: { type: string, pattern: '^[0-9]{6}$', writeOnly: true }
        confirmNewPin: { type: string, pattern: '^[0-9]{6}$', writeOnly: true }
    Account:
      type: object
      required: [accountId, accountNumberMasked, accountType, status, balance, currency, openedAt]
      properties:
        accountId: { $ref: '#/components/schemas/Uuid' }
        accountNumberMasked: { type: string }
        accountType: { type: string, enum: [CHECKING] }
        status: { $ref: '#/components/schemas/AccountStatus' }
        balance: { $ref: '#/components/schemas/MoneyAmount' }
        currency: { $ref: '#/components/schemas/Currency' }
        openedAt: { $ref: '#/components/schemas/Instant' }
...
    OperatorCustomerView:
      type: object
      required: [customerId, fullName, phone, email, isPinSet, createdAt, accounts]
      properties:
        customerId: { $ref: '#/components/schemas/Uuid' }
        fullName: { type: string }
        phone: { $ref: '#/components/schemas/PhoneNumber' }
        email: { $ref: '#/components/schemas/EmailAddress' }
        isPinSet: { type: boolean }
        createdAt: { $ref: '#/components/schemas/Instant' }
        accounts: { type: array, items: { $ref: '#/components/schemas/Account' } }
...
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    Problem:
      type: object
      required: [type, title, status, detail, instance, code, correlationId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { type: string }
        correlationId: { $ref: '#/components/schemas/Uuid' }
        fieldErrors: { type: array, items: { $ref: '#/components/schemas/FieldError' } }
...